extracted mint one shot and fee oracle deploy
