+++ app/mintFeeOracleNftId.ts
@@ -7,12 +7,11 @@ import { tokenName } from "./constants";
 import { getProtocolParams } from "./utils/getProtocolParams";
 import { tryGetMarketplaceConfig } from "./utils/tryGetMarketplaceConfig";
 import { withFolder } from "./utils/withFolder";
+import { getMintOneShotTx } from "./txns/getMintOneShotTx";
 
 async function main()
 {
     const cfg = tryGetMarketplaceConfig();
-
-    const env = cfg.envFolderPath;
 
     const privateKey = cfg.signer.skey;
     const addr = cfg.signer.address;
@@ -20,54 +19,17 @@ async function main()
     const utxos = await koios.address.utxos( addr );
     console.log( utxos.map( u => JSON.stringify( u.toJson(), null, 2 ) ) );
 
-    const utxo = utxos[0];
-    const ref = utxo.utxoRef;
-
-    const feeOracleNftPolicy = makeFeeOracleNftPolicy( PTxOutRef.fromData( pData( ref.toData() ) ) );
-
-    await withFolder( env );
-
-    await cli.utils.writeScript( feeOracleNftPolicy, `${env}/feeOracleNft_${ref}.plutus.json` );
-
-    const policy = feeOracleNftPolicy.hash;
-
-    await writeFile(`${env}/last_ref_used`, ref.toString(), { encoding: "utf-8" });
-    await writeFile(`${env}/feeOracleNft_${ref}.policy`, feeOracleNftPolicy.hash.toString(), { encoding: "utf-8" });
-
     const txBuilder = new TxBuilder(
         await getProtocolParams()
     );
 
-    const mintedValue = new Value([
-        Value.singleAssetEntry(
-            policy,
-            tokenName,
-            1
-        )
-    ]);
+    const utxo = utxos[0];
 
-    const tx = txBuilder.buildSync({
-        inputs: [{ utxo }],
-        collaterals: [ utxo ],
-        collateralReturn: {
-            address: addr,
-            value: Value.sub(
-                utxo.resolved.value,
-                Value.lovelaces( 5_000_000 )
-            )
-        },
-        mints: [
-            {
-                value: mintedValue,
-                script: {
-                    inline: feeOracleNftPolicy,
-                    policyId: policy,
-                    redeemer: new DataI( 0 )
-                }
-            }
-        ],
-        changeAddress: addr
-    });
+    const tx = await getMintOneShotTx(
+        txBuilder,
+        utxo,
+        addr
+    );
 
     tx.signWith( privateKey );
